Switch approach to instead default cameras to writing out background motion vectors, with an opt-out component.
