+++ crates/bevy_core_pipeline/src/skybox/mod.rs
@@ -27,21 +27,14 @@ use bevy_render::{
 use bevy_shader::Shader;
 use bevy_transform::components::Transform;
 use bevy_utils::default;
-use prepass::SkyboxPrepassPipeline;
 
-use crate::{
-    core_3d::CORE_3D_DEPTH_FORMAT, prepass::PreviousViewUniforms,
-    skybox::prepass::init_skybox_prepass_pipeline,
-};
-
-pub mod prepass;
+use crate::core_3d::CORE_3D_DEPTH_FORMAT;
 
 pub struct SkyboxPlugin;
 
 impl Plugin for SkyboxPlugin {
     fn build(&self, app: &mut App) {
         embedded_asset!(app, "skybox.wgsl");
-        embedded_asset!(app, "skybox_prepass.wgsl");
 
         app.add_plugins(UniformComponentPlugin::<SkyboxUniforms>::default());
 
@@ -50,21 +43,13 @@ impl Plugin for SkyboxPlugin {
         };
         render_app
             .init_resource::<SpecializedRenderPipelines<SkyboxPipeline>>()
-            .init_resource::<SpecializedRenderPipelines<SkyboxPrepassPipeline>>()
-            .init_resource::<PreviousViewUniforms>()
             .add_systems(ExtractSchedule, extract_skybox)
-            .add_systems(
-                RenderStartup,
-                (init_skybox_pipeline, init_skybox_prepass_pipeline),
-            )
+            .add_systems(RenderStartup, init_skybox_pipeline)
             .add_systems(
                 Render,
                 (
                     prepare_skybox_pipelines.in_set(RenderSystems::Prepare),
-                    prepass::prepare_skybox_prepass_pipelines.in_set(RenderSystems::Prepare),
                     prepare_skybox_bind_groups.in_set(RenderSystems::PrepareBindGroups),
-                    prepass::prepare_skybox_prepass_bind_groups
-                        .in_set(RenderSystems::PrepareBindGroups),
                 ),
             );
     }
